Rename 'Hover' to 'HoverState'
