+++ packages/dx-chart-core/src/utils/hover-state.js
